requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
einops>=0.6.0
